fix: fix pretrained urls and rename module
