+++ library/src/getitune/backend/lightning/models/classification/hlabel_models/torchvision_model.py
@@ -18,7 +18,7 @@ from getitune.backend.lightning.models.classification.heads import (
 from getitune.backend.lightning.models.classification.hlabel_models.base import LightningHlabelClsModel
 from getitune.backend.lightning.models.classification.losses import AsymmetricAngularLossWithIgnore
 from getitune.backend.lightning.models.classification.necks.gap import GlobalAveragePooling
-from getitune.backend.lightning.models.classification.utils.load_weights import TorchvisionWeightsLoader
+from getitune.backend.lightning.models.classification.utils.pretrained_weights import TorchvisionWeightsLoader
 from getitune.backend.lightning.schedulers import LRSchedulerListCallable
 from getitune.metrics.accuracy import HLabelClsMetricCallable
 from getitune.types.label import HLabelInfo
