+++ library/src/getitune/backend/lightning/models/classification/multiclass_models/vit.py
@@ -21,8 +21,8 @@ from getitune.backend.lightning.models.classification.heads import (
 from getitune.backend.lightning.models.classification.multiclass_models.base import (
     LightningMulticlassClsModel,
 )
-from getitune.backend.lightning.models.classification.utils.load_weights import VisionTransformerWeightsLoader
 from getitune.backend.lightning.models.classification.utils.pretrained_urls import VIT_PRETRAINED_URLS
+from getitune.backend.lightning.models.classification.utils.pretrained_weights import VisionTransformerWeightsLoader
 from getitune.backend.lightning.schedulers import LRSchedulerListCallable
 from getitune.backend.lightning.tools.explain.explain_algo import ViTReciproCAM, feature_vector_fn
 from getitune.data.entity import PredictionBatch, SampleBatch
